Fix dtype mismatch in matrix multiplication for FP16 mixed precision
Fixed RuntimeError: expected mat1 and mat2 to have the same dtype

The issue occurred when using FP16 mixed precision (autocast):
- Image features were in FP16 (from autocast context)
- Text features could be in FP32 (when cached or computed outside autocast)
- Matrix multiplication failed due to dtype mismatch

Solution: Explicitly cast text_features to match the dtype of image features
before matrix multiplication using .to(features.dtype)

Fixed in two locations:
- sclip_segmentor.py:827 - sliding window inference
- sclip_features.py:282 - compute_dense_similarity

This ensures compatibility with FP16 mixed precision while maintaining
numerical stability.

+++ code/models/sclip_features.py
@@ -278,7 +278,8 @@ class SCLIPFeatureExtractor:
         dense_flat = dense_features.reshape(H * W, D)  # (H*W, D)
 
         # Compute similarity: (H*W, D) @ (D, num_classes) = (H*W, num_classes)
-        similarities = dense_flat @ text_features.T
+        # Ensure both tensors have the same dtype (important for FP16 mixed precision)
+        similarities = dense_flat @ text_features.to(dense_flat.dtype).T
 
         # Reshape back to spatial: (num_classes, H, W)
         similarities = similarities.T.reshape(len(texts), H, W)
